+++ src/main/java/org/xandercat/cat/back/CatBack.java
@@ -39,6 +39,7 @@ public class CatBack {
 	
 	public static final String LOG_KEY = "-l";
 	public static final String IMMEDIATE_BACKUP_KEY = "-b";
+	public static final String DRY_RUN_KEY = "-dryrun";
 	
 	private static final Logger log = LogManager.getLogger(CatBack.class);
 	
@@ -48,12 +49,14 @@ public class CatBack {
 		ArgumentProcessor argumentProcessor = new ArgumentProcessor();
 		argumentProcessor.addValidSwitchValuePair(IMMEDIATE_BACKUP_KEY, "<filename>", "Backup the given Backup Profile", null);
 		LoggingConfigurer.addArgumentProcessorSwitchValuePair(argumentProcessor, LOG_KEY);
+		argumentProcessor.addValidSwitch(DRY_RUN_KEY, "Execute backups as simulations without actually updating the file systems.");
 		argumentProcessor.process(args);
 		if (LoggingConfigurer.configureLogging(argumentProcessor, LOG_KEY, Level.INFO, LoggingConfigurer.Target.FILE, "catback.log")) {
 			log.info("Logging configuration updated.  Application will need to be restarted for changes to take effect.");
 		} 
 		PlatformTool.setApplicationNameOnMac(APPLICATION_NAME);
 		final String backupFilename = argumentProcessor.getValueForSwitch(IMMEDIATE_BACKUP_KEY);
+		final boolean dryRun = argumentProcessor.isSwitchPresent(DRY_RUN_KEY);
 
 		// load application settings
 		Object settingsObject = null;
@@ -98,7 +101,7 @@ public class CatBack {
 		SwingUtilities.invokeLater(new Runnable() {
 			public void run() {
 				if (backupFilename == null) {
-					launchUI(settings);
+					launchUI(settings, dryRun);
 				} else {
 					launchImmediateBackup(backupFilename);
 				}
@@ -106,8 +109,11 @@ public class CatBack {
 		}); 
 	}
 
-	private static void launchUI(CatBackSettings settings) {
+	private static void launchUI(CatBackSettings settings, boolean dryRun) {
 		CatBackFrame ui = new CatBackFrame(APPLICATION_NAME, APPLICATION_VERSION, settings);
+		if (dryRun) { 
+			ui.setDryRun(dryRun);
+		}
 		ui.setVisible(true);
 	}