+++ src/main/java/org/xandercat/cat/back/engine/worklet/BackupEngineWorklet.java
@@ -13,6 +13,8 @@ public abstract class BackupEngineWorklet<T> extends SwingWorklet<T, String> {
 
 	public abstract String getTitle();
 	
+	public abstract void enableDryRun(String dryRunPrefix);
+	
 	protected void advanceProgress(long progressBy) {
 		this.progress += progressBy;
 		setProgress(this.progress);